RIA-2493 Prevent AIP events being triggered for Repped journey (#269)
As part of the aip flow we are adding new states and events. The TCW
could trigger those events so we want to prevent this by returning them
an error. Added a new handler that will check the journey type for
certain events and throw an exception to prevent any further handlers
firing. The exception is then handled and an error returned to CCD.