RIA-2541 Added reasons for appeal data
When a case office requests the reasons for appeal they are now
presented with a form to select the required by date. Added code to
populate the form and set the date to 4 weeks in the future. Made
REQUEST_RESONS_FOR_APPEAL a direction so it will be seen in the
directions tab.